refactor(wireframe/connection): use Tokio task-local storage for connection context
- Replace thread-local storage with Tokio task-local storage for ConnectionContext
- Add task-ID keyed registry to mirror context for synchronous app factory use
- Introduce scope_current_context API to manage context scope in async tasks
- Update handshake and tests to use scoped task-local context
- Move handshake BDD tests to separate handshake_bdd.rs module
- Improve async context preservation across Tokio multi-threaded runtime
